Add Magento server client
Add a Magento server client. Magento is unique in that it's not a
single service like Twitter or Tumblr, but instead is a hosted
instance. This requires us to construct with the base uri of the
host.

In addition, Magento has two separate authorization urls based on
whether we are authenticating as an admin or a store user/customer.